GaniProg 0.26
- removed Reposition() in Ingredient.cs

ShopManager
- formerly OrderInventotyUI.cc
- added to MainGameScene
- added prefab spawnpoints
