--- OP_ChopChop!/Assets/OccAssets/OccAssets Prefabs/UI Prefabs/InventoryUIPrefab/NewStoreMenu.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 89c938e25a4032044bb0a20f889b8c3e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
