--- OP_ChopChop!/Assets/_Scripts/Managers/Game Handlers/TEMP_OrderInventoryUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b887e45d41c3bf14cb7cbb4f5faa0857
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
