--- OP_ChopChop!/Assets/_Scripts/Managers/Kitchen Managers/ShopManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e5d428f1f9f8b0a4ca895160cbec3e4e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
